Moved the brushing tests into a separate test class (CDSVisualizationBrushingTest). Removed unused variables from CDSVisualizationTest. Replaced waitForElement(spinner) with a sleep.
